requirements.txt
asgiref==3.5.0
Django==3.2.12
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
dynaconf==3.1.7
future==0.18.2
importlib-metadata==4.11.1
Markdown==3.3.6
pytz==2021.3
serializers==0.2.4
sqlparse==0.4.2
zipp==3.7.0
